codesniffer: Prepare for adding missing WordPress rules
We're currently pulling in rules from WordPress-Core, WordPress-Docs,
and WordPress-Extra. It turns out that that misses four rules from
"WordPress":

- WordPress.DB.DirectDatabaseQuery
- WordPress.DB.SlowDBQuery
- WordPress.Security.ValidatedSanitizedInput
- WordPress.Utils.I18nTextDomainFixer

The last of those is no problem, it's redundant to WordPress.WP.I18n
(for some reason they put the fixing logic into a separate rule) and
doesn't report anything unless configured anyway.

For the other three we'll have to clean up our existing code to fix at
least the 224, 35, and 808 instances in the non-excluded files before we
can enable them. So for now we'll explicitly disable them, then enable
them as we fix them.