Ignore all hits on `WordPress.DB.SlowDBQuery`
Several are readily identifiable as false positives. The rest I figure
we've been running these queries in production for long enough that if
they were really a problem we'd already have issues filed about them.

Overall I can't say I'm terribly impressed with this sniff. It just
blindly looks for assignments to certain array keys with no context. And
thanks to WordPress's general lack-of-abstraction DB layer I saw plenty
of places where it missed uses of the things it's sniffing for. If it
weren't for wpcom enforcing it, I'd not bother with it in our ruleset.