PHPCS: Fix WordPress.Security.ValidatedSanitizedInput outside of Jetpack and WAF
Mostly this means adding `wp_unslash()` and `filter_var()` calls, and
sometimes `isset()` checks too. Occasionally it means ignoring the sniff
in code that's already doing better than `filter_var()` would do.

Jetpack is skipped for now because it's big. WAF is skipped for now
because it's not immediately clear to me whether WordPress's
`wp_unslash()` is loaded at the time it runs.