Make subscribe button work in email & notifications
The subscribe button used in premium content blocks should open a dialog
to pay for the premium content when clicked, however when the subscribe
button was displayed in emails or in wp.com notifications it didn't do
that. It did work as intended when a recurring payment button was used
directly.

This was happening because the underlying jetpack/button wasn't using
the correct `url` - none was set.

This has been addressed by:
 1. Changing the template in premium-content to specify the
    `passThroughAttributes` on the jetpack/button
 2. Updating recurring-payments so that if the planId prop is changed,
    it sets the url and uniqueId props too.

Fixes #23696