PHPCS: Fix WordPress.Security.ValidatedSanitizedInput outside of Jetpack and WAF (#23942)
* codesniffer: Prepare for adding missing WordPress rules

We're currently pulling in rules from WordPress-Core, WordPress-Docs,
and WordPress-Extra. It turns out that that misses four rules from
"WordPress":

- WordPress.DB.DirectDatabaseQuery
- WordPress.DB.SlowDBQuery
- WordPress.Security.ValidatedSanitizedInput
- WordPress.Utils.I18nTextDomainFixer

The last of those is no problem, it's redundant to WordPress.WP.I18n
(for some reason they put the fixing logic into a separate rule) and
doesn't report anything unless configured anyway.

For the other three we'll have to clean up our existing code to fix at
least the 224, 35, and 808 instances in the non-excluded files before we
can enable them. So for now we'll explicitly disable them, then enable
them as we fix them.

* Drop WordPress.DB.SlowDBQuery

* PHPCS: Fix WordPress.Security.ValidatedSanitizedInput outside of Jetpack and WAF

Mostly this means adding `wp_unslash()` and `filter_var()` calls, and
sometimes `isset()` checks too. Occasionally it means ignoring the sniff
in code that's already doing better than `filter_var()` would do.

Jetpack is skipped for now because it's big. WAF is skipped for now
because it's not immediately clear to me whether WordPress's
`wp_unslash()` is loaded at the time it runs.

* That wasn't supposed to be here

* device-detection: Work around possible non-presence of `wp_unslash`

* Add stubs for wp_unslash etc in packages/status, plugins/boost

* Add one more ignore