Publicize: Move the LinkedIn reauth check to within the loop
When we check the status of the Publicize connections, there is
currently a check outside of the loop for whether a LinkedIn connection
needs to be reauthorised, but in testing another change, I was seeing
warnings about variables being undefined.

Looking at the intention of the code, this check should be moved within
the loop, so it is performed on each connection. This change does that.