jetpack: Remove create-react-class and react-pure-render deps (#24055)
These are back-compat packages for code from before React 16. Let's fix
the uses to use modern syntax instead.

This also deletes mixins/url-search. Modern React doesn't use mixins,
and this was only used in one place so it's easier to just inline it
rather than converting it into a wrapper component or something.