Replace uses of `pnpx` with `pnpm` (#24210)
In pnpm 7.0, the `pnpx` command is changing behavior to only do
"download to a temporary environment and run", it no longer runs
already-installed binaries.

Instead, `pnpm` or `pnpm exec` should be used. Or, when running from
package.json scripts, we can omit it entirely since `node_modules/.bin/`
is already in the path.