Publicize: Add a Jetpack Publicize connections endpoint.
This adds a new endpoint, that proxies a request to WPCOM in order to
retrive the Publicize connections. It also makes a change to the initial
state, so that on WPCOM we call the API directly. It might be better to
just have an `isWPCOM` boolean so that it's more useful.