webpack-config: Add a plugin to make module ids more deterministic
In pnpm, the paths to packages in node_modules include the package's
version number. This winds up in the module ID used by webpack, which
causes the "deterministic" IDs it assigns in production mode to change
whenever a package is updated (it also changes the named IDs in
development mode, but we don't really care about that).

This new plugin is a simple modification of Webpack's deterministic ID
plugin that strips the version numbers from the identifier before
assigning the numeric ID.