jetpack: Use browser URL API instead of polyfills (#24234)
Editors for calendly and opentable were using `url-polyfill` to polyfill
the browser URL API. But it looks like all browsers we support already
supply it natively, so let's drop the polyfill.

Donations, Stripe connection notification, and Tiled Gallery were using
a polyfill for the Node URL API, which Node itself seems to be in the
process of deprecating. Again, drop that and use the browser URL API
instead.