ranovate: Log GitHub API rate limits pre- and post-run
Yesterday a run failed because the token hit the 5000/hr rate limit.
Let's track the rate limit on each run to see whether it seems the token
is chronically near the limit or if that was a fluke.

Note hitting the /rate_limit endpoint doesn't itself count towards the
rate limit.