Codesniffer: allow the use of lowercase WordPress
While the rule is important, we currently attempt to autofix PHPCS errors in our pre-commit hook. When this happens and an array key is automatically renamed from wordpress to WordPress for example, we may inadvertently add errors via the automated tools.

Let's not do that.

Related discussion: p1652454457615319-slack-CBG1CP4EN