Remove pnpm hoisting for eslint (#24355)
This involves moving the bits of eslint configuration that load plugins
or sharable configs into `tools/js-tools/eslintrc/`, so the plugins and
configs can be loaded relative to that directory instead of relative to
wherever the origin `.eslintrc.js` file happens to be.

This also drops the dep on `@sveltejs/eslint-config`, which hasn't been
used since #23482.

And for some reason projects/js-packages/shared-extension-utils/ had
both .eslintrc.js and .eslintrc.cjs, with identical content. So I
deleted one.