ci: Set codecov upload token
While codecov's documentation is clear that the token is not required
for public repositories, it turns out that, without the token, it uses a
GitHub API token that's shared with everyone else in the world.

Since [they say][1] that using the token (in combination with their
GitHub app) will avoid that rate limiting, let's try that.

[1]: https://href.li?https://github.com/codecov/uploader/issues/740#issuecomment-1130379318