Social: Create the Jetpack Social version of the Publicize panel (#24466)
* [not verified] Publicize Components: Move remaining components to the package.

This PR builds on #24408 and moves over the remaining components to the
package which are required by Jetpack Social.

This is mainly moving files, but with a few changes:

- JSDoc has been updated to satisfy the new lint rules
- Styles have been moved into the components where possible
- We no longer display the message box and twitter options when the
  toggle is used to disable Publicize

* Update the lock file

* Update the connections URL until the Jetpack Cloud page is ready.

* [not verified] Social: Create the Jetpack Social version of the Publicize panel

Jetpack Social needs its own implementation of the Publicize panel,
displayed in the editor sidebar. This is because we're not handling paid
plans yet, and the paid features (Republicize) aren't available. When we
do, it's unlikely that we'll handle everything in the same way as the
Jetpack plugin.

* Set the connections link to the Jetpack redirect

* Update project versions

* Add the description to the panel.

* Disable the editor extension is Publicize is disabled.

* Simplify the Description component definition.