Publicize: Replace the initial settings form in the Classic Editor
In the classic editor when there is no Publicize connections, we
currently display some links to authenticate directly with the networks,
but they redirect back to a deprecated options page which only exists
within Jetpack.

This change replaces that form with the Jetpack redirect link that will
take the user to the appropriate connections page on Calypso/Jetpack
Cloud.