Social: Prevent the editor extension loading with Jetpack active (#24521)
The rest of the plugin is compatible with Jetpack. The module can be
toggled from both plugins, and the Publicize is only loaded once, so the
Classic editor works correctly.

But we were loading the editor extension, so with both plugins you have
both sidebars. This change prevents the Social one from loading when
Jetpack is active.