tooling: Improve version tag handling
`tools/deploy-to-svn.sh` will now handle tags with a "v" prefix.

The autotagger action can now be configured not to generate the "v"
prefix. The prefix is still the default.

The autorelease action will now accept tags without the "v" prefix, and
with 2 or 4 version components. The matching is also more strict so it
shouldn't ever fail due to running on an invalid tag.

The (still unused) wp-svn-autopublish action's tag matching was made
more strict too.

Note the npmjs-autopublish action still requires a "v" prefix and
exactly 3 components.