git: pre-commit hook shouldn't check files twice
We had been running non-excluded files through both `phpcs`/`eslint` and
`phpcs-changed`/`eslint-changed`. If we do the former there's no need
for the latter.

For that matter, `eslint --fix` will report any remaining problems and
exit with a proper code, so there's no need to run a file through both
`eslint --fix` and `eslint`. Unfortunately `phpcbf` isn't so helpful, so
we still have to run through `phpcs` after that one.