eslint-changed: Change to esm and use jest for tests (#24626)
The ESM switch was very straightforward, all that was needed was to swap
`require` for `import` and replace some uses of `__dirname` in the tests.

The switch to `jest` was a little more complicated:

* The first part was simple enough, once `--experimental-vm-modules` was
  added then it worked without even disabling code transforms.
* But jest's resolver choked on a fancy feature in the new version of
  chalk, so a workaround had to be added for that. Sigh.
* And then getting coverage working required a restructure of all the
  code, because jest won't record coverage for code in child processes.

And then when I was getting all the tests to work with the restructured
code, I noticed that "Filters by filename" was testing for incorrect
behavior (it wasn't actually filtering). So I had to fix that too.