cli: Use jest for tests (#24635)
We're trying to standardize on jest across the monorepo, so replace use
of mocha+chai+sinon with jest for the CLI's tests.

While updating the tests, I also found that `promptForProject` was
returning an error instead of throwing one like it should. You could hit
this with `jetpack changelog version plugins/foobar current` and
`jetpack watch plugins/foobar`.