eslint: Enable more jest linting
This adds more linting rules for jest tests, to help make things more
consistent and to provide better messages on failures.

Of the added rules, the only one of note that needed fixes anywhere was
one preferring `toBe()` over `toEqual()` for primitive values, which
generates better messages on failure.

Also add linting rules for `@testing-library/react`.

Of the added rules, the only one that needed fixes anywhere was in an
Instant Search test where it wanted us to use `screen.getByRole()`
instead of calling `getByRole()` on the result of `render()`.