Fix bug preventing images in payment buttons
Inline images didn't work in payment buttons - the image wasn't rendered
at all, and neither was any accompanying text.

This was happening due to a small bug introduced in #23487 where the
regular expression wasn't just replacing the href attribute and the "
surrounding it's values, but all of the content between the " opening
the href value and the last " in the user-supplied content.

This commit fixes it by using an ungreedy regex, which should mean just
the href attribute and it's delimiters are matched.

Fixes #24662