codesniffer: Fix dev dep on WPCS (#24711)
For a while now we've been depending on a particular commit of WPCS that
works with PHP 8. We had been doing that as `dev-develop#<sha>`, which
Composer only halfway supports.

Now that WPCS has updated their dev-develop to depend on phpcs 3.7,
Composer is getting confused because it's still trying to apply the
composer.json metadata from the tip of dev-develop rather than the sha
we're pointing to.

Since we already have a fork from #24627, let's go ahead and use that
here too to avoid the problem. Hopefully someday WPCS will finally do a
release...