[RNMobile] Prevent validation error when viewing VideoPress markup within app (#24548)
Currently, for the web, an editor.js file "intercepts" the video block
to make some modifications when saved. With this PR, a native version of
the editor.js file has been created to enable us to make our own
native-specific modifications.

The new file defines attributes and utilises VideoPress' current save.js
file to prevent a validation error when viewing a block with VideoPress
markup within the app.