photon: Do not return Photonized URLs to the block editor in WP 6.0 (#24769)
In WordPress 6.0 (or corresponding versions of Gutenberg), media
requests started using `context=view` instead of `context=edit`. Which
makes sense, since the image is being viewed, not edited. But we were
using that context in Photon to tell when the request was coming from
the editor so we could avoid having the photonized URLs saved in the
post content.

To identify requests from the editor now, we register a middleware with
`@wordpress/api-fetch` from our shared editor setup JS to add an HTTP
header on all requests, and have the Photon class check for that header.