js/idc: Remove unused testing infrastructure
When I went to convert this package to use jest instead of
js-test-runner, I found no tests to convert. So I left it at setting up
test linting and removed the rest.

If someone wants to add tests to the package in the future, they can
copy-paste from one of the other packages at that time, when they'll
have actual tests to know what specifically needs to be set up.