webpack-config: Remove unused testing infrastructure
This lets us drop the last use of `nyc`, and remove 19 other
dependencies it brings in.

If someone wants to add tests to the package in the future, they can
copy-paste from one of the other packages at that time, when they'll
have actual tests to know what specifically needs to be set up.