Fix double-quotes handling in subscription button (#24763)
* [not verified] Fix double-quotes handling for the subscription block button

When the subscription block is saved, the submitButtonText attribute
has a double-quote encoded as a Unicode character. However, the submit_button_text
shortcode attribute is not encoded, and it breaks button text when double-quotes
are used.

This PR adds HTML encoding for double-quotes in submit_button_text. HTML encoding
is used as this text is later directly used in the generated HTML code.

* [not verified] changelog

* [not verified] Optimize location for double-quotes escaping

Code that handles special characters escaping is extracted
to the separate function. Then, escaping for additonal
characters is added.

* [not verified] Regenerate fixtures for the test that includes single-quotes

* [not verified] Add escaping for square brackets, nbsp, and zwsp

* [not verified] Remove explicit list of escaped characters from the comment.

* [not verified] Extract encoding function to the separate file

Extract the function to the separate utlity file to avoid
redefining it on every render.