live-branches: Allow for not installing Jetpack (#24843)
Jurassic Ninja installs the release version of the Jetpack plugin by
default, unless a `nojetpack` option is supplied. Let's add a UI for
that to simplify testing of the standalone plugins without Jetpack
itself installed.

To keep the UI simple, we add "Jetpack" in the existing "Plugins" list,
pre-checked to preserve the current behavior as the default. In the JS
we invert the sense of this checkbox so `nojetpack` is added to the URL
when the checkbox is unchecked.