Repo Gardening: introduce new task to triage to projects
This should allow us to add issues to specific GitHub Project boards based on the labels used in the issue.

This only supports V2 Project boards for now. Classic GitHub projects are apparently going to be deprecated soon, and the existing boards will be migrated to new Projects (currently in Beta). Hence why I went with something that supports the new type of Projects.

Internal reference: pciE2j-18N-p2