[not verified] Inline BlockVariationPicker in PaymentsIntro
The PaymentsIntro block used the experimental BlockVariationPicker from
Core Gutenberg [0] in order to let people select a payment block.

This change copies some of the existing BlockVariationPicker logic from
core Gutenberg into the PaymentsIntro block itself. The main reason for
this is to allow greater UI flexibility, specifically integrating block
patterns into the PaymentsIntro block (see #23422).

Additionally, the BlockVariationPicker is experimental and subject to
drastic and breaking change, it was arguably a misuse of the component -
 the blocks presented are not variations of each other, and are
technically not really related at all.

[0]
https://github.com/WordPress/gutenberg/blob/93ab227ef76a1a73720582b6958cf1a079a91aa8/packages/block-editor/src/components/block-variation-picker/README.md