jetpack-live-branches: Improve behavior when switching tabs in GH (#24977)
GitHub doesn't always reload the page when switching between tabs on a
PR. Use MutationObserver to watch for that and re-add the UI when
needed. This also allows us to preserve the UI when the PR's description
is edited.

Also, avoid showing the UI on pages like the "details" for Codecov's
checks and avoid breaking the notification settings.