js/analytics: Avoid breaking if `_superProps` hasn't been set
When #24759 changed from `lodash.assign` to `Object.assign`, a small
difference in behavior was unnoticed: the former when passed undefined
as the first parameter will create a new object to assign into, while
the latter just throws an error.

One call was counting on the lodash behavior. This fixes it.