Replace `softprops/turnstyle` with a local rewrite (#25097)
We recently had an issue where turnstyle was allowing jobs to run in
parallel, which proved impossible to debug as that action does not log
any details of the API calls.

Also upstream seems unresponsive to PRs.

So, let's reimplement this locally. Not planning on releasing it
publicly at this time, but we could if someone wants it.