Repo Gardening: create a new "Reply to customers Reminder" task
This new task will run when issues get closed.

It will check for a few things:
- Is the issue important (has a High or Blocker priority label)?
- Does it have a comment listing support references, where more than 10 answered (unchecked) support references are listed?

If it finds such a comment, it will attempt to post a message in a Slack channel (specified in a GitHub secret, so it can be different for each repo), asking the triage team to reply to all those tickets.

Internal reference: pcLjiI-Ro-p2