Updated shortNumberFormat function
I have set up the shortNumberFormat function and created a readme file for the function.

I tested the function separately using hard coded regions, but I am new to development and developing for Jetpack and don't know how to test this in the context of the plugin. It should work in the same way as numberFormat.