Gather support references: add option to send a Slack message
When issuees start gathering a lot of tickets (at least 10), let's send a Slack message so the issue can be escalated if needed.

We only want to send that Slack message so after doing it, we add a note to the comment that lists all support references to indicate it was escalated. Later, we search for that note before we try to send a new one.
Internal reference: pciE2j-1iy-p2