Make default payment block plan names dynamic
When adding a new plan, the plan name will update depending upon the
payment options selected - whether it's a donation  and what the payment
frequency is.

This aids users in naming their plans appropriately and reduces the
likelihood of having umpteen plans called 'Monthly subscription'
when they're nothing of the sort.

If the user explicitly chooses a name then that is respected and not
overwritten.