Adapt premium block nudge on constrained widths
A nudge to upgrade plan is displayed for Premium Blocks, however on
constrained widths, such as when displayed inside a column block, this
nudge is not displayed correctly. The nudge becomes hard to read and is
obscured by the toolbar menu.

The toolbar is displayed in relation to the block element, this change
attempts to capitalise on that by setting the style of the block element
in javascript to leave enough space for the upgrade nudge. It does not
work because this is calculated too late - after the toolbar rendering
code has ran.