Stats: don't show links in dashboard widget when module inactive
The Jetpack Stats dashboard widget, while named that way, is used for more than just stats. When the stats module is inactive, we still register the widget and use it to display other types of stats: Akismet and Protect.

In those situations, the stats config links are not that useful and should be removed.