Actions: Don't notify Slack for issue gardening failures
Now that we're running the Gardening action on issues as well as PRs and
merges, we find that the Slack notification if an issue gardening run
fails complains about the head commit on trunk.

After some discussion (p1660745332014549/1660736016.033519-slack-C034JEXD1RD)
we've decided that we don't need Slack notifications for issue gardening
failures, so let's disable them.