Fix GitHub mirror repo actions (#25782)
* Actions: Fix GitHub filter patterns

WTF, GitHub? Why "Ranges can only include `a-z`, `A-Z`, and `0-9`"?

We already have code in the called script to re-check the ref, so just
loosen the pattern.

* autorelease action: Handle `.extra.autorelease` being boolean

Boolean is supposed to be valid, but `jq` was erroring when trying to
access a property on it. Use the `.foo?` syntax to fix that.