phpcs: Re-enable WordPress.DB.DirectDatabaseQuery sniff
We disabled this during the big phpcs cleanup project because we didn't
need to or want to fix it at the time. That project is done now, and the
rule is thought to be useful enough to not permanently disable, so we
should re-enable it so new instances don't get added.