Revert "build: Move "Test plugin upgrades" to a post-build workflow (#25892)"
This reverts commit 4883758d937ec80850e417f4ffb413db7a0a24f4.

It looks like GitHub has a fairly critical bug (which they've apparently
not cared about for 2 years[1]) where the check run created from Actions
gets assigned to a random suite. If that random suite happens to be one
of the Gardening runs, it can get lost when the next Gardening run
happens.

[1]: https://github.com/orgs/community/discussions/24616