renovate: Fix ignorePaths config
The default ignorePaths ignores all files in `/tests/` directories,
which includes our `projects/plugins/*/tests/e2e/package.json`.

It turns out adjusting this configuration can't be done from the global
config in renovate-config.js, as the preset expansion overwrites it no
matter what I try. But we should be able to do it by adding back a
`renovate.json5` repo-level config file.

Most of the configuration stays in the global config, because that can
do things like read the list of all packages.

This also adds `.json5` as a recognized extension for the prettier
invocations, since it wasn't there already.