tools: Improve check-intra-monorepo-deps version handling
The old check was kind of dumb: if the existing dep started with `^`
then it would do like "^1.2", otherwise it would do like "1.2.x-dev".
Except for packages, where it would always do like "^1.2".

That broke when used on a plugin release branch where someone had pinned
a dep like "1.2.3".

The new behavior is a bit more accurate: for plugins, it it ends with
"-dev" then it does like "1.2.x-dev", otherwise it does like "^1.2.3"
(matching the recently-changed behavior of tools/version-packages.sh).
For packages it always does like "^1.2", and for everything else it
always does like "1.2.x-dev".

Plus, to remove the need for a pinned "1.2.3" dep, when doing the
composer.lock update it will force a "^1.2.3" dep to install that
specific version rather than allowing a newer version to be pulled in.
With one caveat: indirect deps might still be updated incorrectly.
Humans will have to watch for that, at least until a different PR adds a
CI test for it.