autoloader: Skip tests with incompatible PHP and Composer versions
We have some tests here that run with old versions of Composer, so old
that they don't work with PHP 8.1.

This adds some metadata to the composer version list to indicate which
PHP versions it's compatible with, and skips the test if we don't know
of any Composer version compatible with both the autoloader version and
the PHP version.