Fix type error for image array (#26110)
By default, top posts set the fallback gravatar for the posts that do not
have a featured image.

There is a filter hook `jetpack_top_posts_widget_image_options` that can be
used to disable the fallback image. Disabling fallback images using the
above hook should not display fallback images for the posts without featured
images.

Instead, it was raising a type error in PHP 8.0, and displaying a broken
image in earlier versions.

This adds checks for no image being returned and omits the image display in
that case.