tools: Refactor `--with` logic into composer-update-monorepo.sh
In #25989 we added logic to version-packages.sh to use `--with` when
calling composer-update-monorepo.sh to force monorepo deps in the
"^1.2.3" style to install version 1.2.3 exactly, rather than picking up
any later version, to better work in release branches where that style
of dep is used.

And in #26072, we did the same thing for check-intra-monorepo-deps.sh
for the same reason.

Chances are any other caller of composer-update-monorepo.sh is going to
want the same too, so let's just move the logic there.