Set Jetpack available_modules to prevent jetpack/modules directory scan
Module tasks related to activation usually check if some module is available before doing the action. This is done by scanning the jetpack/modules directory or by fetching the data from the Jetpack available_modules option. Given that we are mocking Jetpack, there is no modules dir, so we need to enforce a list of available modules by using the option value.