Include mock class to extend Module_Product before testing it
The test-module-product suite is used to test the Module_Product class, that is a class to represent a product that is, in fact, a module inside Jetpack.

The Module_Product class is abstract, so we can't instantiate it to test.

When the test suite got created, we had the Videopress product as a module inside Jetpack, so Videopress was used for the tests. After #25954, Videopress is now an Hybrid product, extending Hybrid_Product class, and therefore we can't use it to teste Module_Product class.

We have two other products that are currently extending Module_Product, but they are overriding some methods and changing the behavior, so they are not good candidates for using in the test suite as well.

To solve this, I am adding this Mock_Module_Product as a pure Module_Product implementation, so it can be the subject of the tests. It just sets the basic names for the product and implements only the abstract methods from the parent class, relying on sample or empty data for the returned values.