My Jetpack: Re-enable My Jetpack module-product tests (#26227)
* Add is_connection_ready method to the Jetpack class mock

* Remove skip command on my-jetpack tests for module-product.php

* Set required Jetpack constants on the Jetpack plugin mock

* changelog

* Set Jetpack available_modules to prevent jetpack/modules directory scan

Module tasks related to activation usually check if some module is available before doing the action. This is done by scanning the jetpack/modules directory or by fetching the data from the Jetpack available_modules option. Given that we are mocking Jetpack, there is no modules dir, so we need to enforce a list of available modules by using the option value.

* Include mock class to extend Module_Product before testing it

The test-module-product suite is used to test the Module_Product class, that is a class to represent a product that is, in fact, a module inside Jetpack.

The Module_Product class is abstract, so we can't instantiate it to test.

When the test suite got created, we had the Videopress product as a module inside Jetpack, so Videopress was used for the tests. After #25954, Videopress is now an Hybrid product, extending Hybrid_Product class, and therefore we can't use it to teste Module_Product class.

We have two other products that are currently extending Module_Product, but they are overriding some methods and changing the behavior, so they are not good candidates for using in the test suite as well.

To solve this, I am adding this Mock_Module_Product as a pure Module_Product implementation, so it can be the subject of the tests. It just sets the basic names for the product and implements only the abstract methods from the parent class, relying on sample or empty data for the returned values.

* Use Mock_Module_Product as the subject for the Module_Product testing

* Update changelog entry

* Rename testing class to use Sample instead of Mock

The class is not actually a mock, but a valid, simple implementation of a Module_Product, so I changed it to Sample_Module_Product for better semantics.