[not verified] Fix upgrade nudges in editor
The upgrade nudges were appearing beneath the block toolbar menu. This
was because the outer div had a display:flex rule thanks to the is-layout-grid
class.

The layouts work at edit time by applying the is-layout-flex css class
to the BlockListBlock (i.e. the component rendered on canvas) and at
view time by adding the same class name to the outermost rendered content.

This commit changes the edit component to go back to <div { ...blockProps }>
but to delete blockProps.className first. This way the `display:flex` won't be
applied in the wrong place.