autotagger: Only tag from release branches if set
If a mirror has a release branch configured, only auto-tag from that
release branch, not trunk.

When we're using a release branch, it probably means it's a plugin that
will need to have had `tools/version-packages.sh` run for a correct
release, meaning that tagging from trunk won't work right.