ci: Set up test env with built version of WordPress (#26389)
WordPress core recently made a change that breaks testing unless
WordPress has been built, and put in a special case only for _their_
unit tests.

The most future-proof option seems likely to be to download a built
version of WordPress (from the "core" repo) to test against, replacing
the unbuilt code from the "develop" repo. But we still need the test
base classes and bootstrapping code from the "develop" repo too.

Rejected alternatives:
* Build the develop repo. They require different software versions, not
  worth messing with.
* Work around the immediate issue. There will probably be others down
  the road. Since the documented way to integration-test plugins
  involves fetching from the "core" repo, that's probably all they'll
  support.

Also, we now have to mock the update API response in a sync test since core
now sees it has the latest version and won't indicate an update, which the
test was relying on.