jetpack: Prepare for de-Fusioning masterbar by removing wpcom checks
Before we can de-Fusion the masterbar PHP code, we need the Jetpack code
to load the Jetpack copy of the CSS instead of trying to load from the
Fusioned paths.

The Fusioned copy of this code will need to continue loading from the
wpcom paths, of course, so the Fusion diff for this PR will not be
merged.