Payments: Make payments-intro pattern filtering more robust (#26465)
When the payments-intro block was filtering patterns to display in it's
pattern selector, it assumed that all patterns have a categories
property. This isn't necessarily the case, and if one pattern didn't,
the whole block didn't work and gave the 'encountered an error' message.

This has been fixed by checking that the property is defined before
using it.