Fix display of wpcom nudges
WPcom nudges were including the nudge as part of the button grid. This
change addresses that by adding HTML data attributes and classnames,
along with a div, so that the wpcom nudge code uses the div controlled
by this block, rather than by introducing its own, where it guesses is
best.