Social: Require a user connection to use the plugin.
If only a site connection is present (for example, when Jetpack Boost
has been set up), then it was causing errors as the editor was loaded.
This was down to us trying to retrieve and test the status of, social
media connections for a non-existent user.

This change adds some extra logic to ensure that we have a user
connection present before we load the editor extension or the main admin
page.