jetpack: Fix Jetpack_PostImages::get_image return value
Historically, the function has been declared as returning an array, but
returned the empty string on failure to find an image.

In #26110 this was changed to return an empty array. Which mostly
worked, but one caller was assuming that any array return was a success.

Let's change it to return null on failure instead. That will still work
with most of the callers that check for `empty()`, as well as the one
checking for a falsey value, as well as this one that checks for
`is_array()`. Overall a null or false return probably makes more sense
semantically too.